+++ src/lapidary/runtime/operation.py
@@ -1,11 +1,10 @@
 import dataclasses as dc
 import functools as ft
-from collections.abc import Callable, Iterable
+from collections.abc import Callable
 
 import typing_extensions as typing
 
 from .model.op import mk_exchange_fn
-from .types_ import SecurityRequirements
 
 OperationMethod = typing.TypeVar('OperationMethod', bound=typing.Callable)
 SimpleDecorator: typing.TypeAlias = Callable[[OperationMethod], OperationMethod]
@@ -15,7 +14,7 @@ SimpleDecorator: typing.TypeAlias = Callable[[OperationMethod], OperationMethod]
 class Operation:
     method: str
     path: str
-    security: Iterable[SecurityRequirements] | None = None
+    security: None = None  # deprecated, not used, TODO remove
 
     def __call__(self, fn: OperationMethod) -> OperationMethod:
         exchange_fn = mk_exchange_fn(fn, self)
@@ -23,7 +22,7 @@ class Operation:
 
 
 class MethodProto(typing.Protocol):
-    def __call__(self, path: str, security: Iterable[SecurityRequirements] | None = None) -> typing.Callable:
+    def __call__(self, path: str) -> typing.Callable:
         pass